program/requirements.txt
pyserial
pyttsx3
